Modal Popup Component 1.
TODO: work on styling ... incorporate design into
